requirements.txt
langchain
langchain-community
langchain
langchain-community
langchain-core
pymupdf
sentence-transformers
langchain-ollama
uvicorn
fastapi
python-multipart
groq
langchain-groq
python-dotenv
langchain_huggingface
dotenv
boto3
tqdm
assemblyai
google-auth
requests
psycopg2-binary
sqlalchemy
pgvector